+++ baseapp/state/state.go
@@ -11,7 +11,7 @@ import (
 )
 
 type State struct {
-	MultiStore storetypes.CacheMultiStore
+	MultiStore storetypes.CommitBranch
 
 	mtx sync.RWMutex
 	ctx sdk.Context
@@ -19,7 +19,7 @@ type State struct {
 	span trace.Span
 }
 
-func NewState(ctx sdk.Context, ms storetypes.CacheMultiStore) *State {
+func NewState(ctx sdk.Context, ms storetypes.CommitBranch) *State {
 	return &State{
 		MultiStore: ms,
 		ctx:        ctx,
@@ -28,7 +28,7 @@ func NewState(ctx sdk.Context, ms storetypes.CacheMultiStore) *State {
 
 // CacheMultiStore calls and returns a CacheMultiStore on the state's underlying
 // CacheMultiStore.
-func (st *State) CacheMultiStore() storetypes.CacheMultiStore {
+func (st *State) CacheMultiStore() storetypes.MultiStore {
 	return st.MultiStore.CacheMultiStore()
 }